Zookeeper log level added and updated curator version (#1831)
* Prepare for next version

* Noting 1.3.0 release date

* zookeeper log level added
