+++ serving-api/src/main/resources/logback.xml
@@ -34,7 +34,7 @@
     <logger name="org.apache.hadoop" level="ERROR"/>
     <logger name="org.apache.spark" level="ERROR"/>
     <logger name="com.stratio.sparta" level="INFO"/>
-    <logger name="org.apache.zookeeper.ClientCnxn" level="OFF" />
+    <logger name="org.apache.zookeeper.ClientCnxn" level="ERROR" />
 
     <root level="ERROR">
         <appender-ref ref="STDOUT" />
